refactor(auth): bearer(optionalEnv()) replaces inferBearer — no host inference
PR review flagged that inferBearer's host heuristic could attach a real
credential to the wrong host: it keyed the env var on the first DNS label
(api.github.com -> GITHUB_TOKEN), so an attacker-influenced or look-alike host
(github.evil.com, api.github.com.evil.com) sharing that label would harvest the
token. Drop host inference entirely and require an explicit env-var name.

Replace the inferBearer strategy with a composition that separates header
creation from secret resolution:

- optionalEnv(name): a resolver beside env()/secretsFile() that resolves a
  value, or *absent* — it never throws (an exported-but-empty var counts as
  absent). Browser-safe for free: no process env => absent.
- bearer(token: Secret | OptionalSecret): for an optional secret, attaches the
  header only when present, else skips it and emits an info event naming the
  var (never the token). The required Secret path is byte-for-byte unchanged
  (still throws via env() on a missing var).

OptionalSecret is branded and type-scoped to bearer, so apiKey/basic/oauth2
cannot be made optional by accident. Removes the now-unused hasEnv() util.
Swaps the inferBearer spec for optional-env.spec.ts; docs updated.
